move data/ to lib/data
