+++ components/hero-unit.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 import Image from 'next/image'
-import { formatURL } from '../data'
+import { formatURL } from '../lib/data'
 import styles from './hero-unit.module.scss'
 import links from '../styles/links.module.scss'
