+++ templates/contact.tsx
@@ -4,14 +4,14 @@ import {
   HeroUnit
 } from '../components/hero-unit'
 import Layout from '../components/layout'
-import { formatPhoneNumber } from '../data'
+import { formatPhoneNumber } from '../lib/data'
 import styles from './contact.module.scss'
 import Container from 'react-bootstrap/Container'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import ContactForm from '../components/forms/contact'
 import { PageProps } from '../pages/[slug]'
-import { WordPressPage } from '../data/types'
+import { WordPressPage } from '../lib/data/types'
 import PageSEO from '../components/seo'
 
 interface ContactPage extends WordPressPage {
