+++ templates/videos.tsx
@@ -9,7 +9,7 @@ import Container from 'react-bootstrap/Container'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import { PageProps } from '../pages/[slug]'
-import { WordPressPage } from '../data/types'
+import { WordPressPage } from '../lib/data/types'
 import PageSEO from '../components/seo'
 
 interface VideoPage extends WordPressPage {
